simplify to single data model, rename xf_to_array -> df_to_xf
